requirements.txt
flask
flask-pymongo
python-dotenv
google-genai
flask-bcrypt
flask-login
datetime
aiohttp